pyproject.toml
[project]
name = "mcp-on-streamlit"
version = "0.1.0"
description = "MCP Client on Streamlit Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.55.0",
    "fastmcp>=2.9.1",
    "langchain-anthropic>=0.3.16",
    "langchain-google-genai>=2.1.6",
    "mcp>=1.10.1",
    "mcp-use>=1.3.3",
    "nest-asyncio>=1.6.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
